add new mysql tests
